+++ types/beatmap/shared/custom/chroma.ts
@@ -1,5 +1,3 @@
-import { IInfoSettingsCustomData } from './heck.ts';
-
 export type LookupMethod = 'Regex' | 'Exact' | 'Contains' | 'StartsWith' | 'EndsWith';
 
 export type GeometryType =
@@ -93,16 +91,3 @@ export type EnvironmentMaterial =
    | 'BillieWater'
    | 'InterscopeConcrete'
    | 'InterscopeCar';
-
-/** Chroma interface for Difficulty Info Custom Data. */
-export interface IChromaInfoCustomData extends IInfoSettingsCustomData {
-   _settings?: {
-      _chroma?: {
-         _disableChromaEvents?: boolean;
-         _disableEnvironmentEnhancements?: boolean;
-         _disableNoteColoring?: boolean;
-         _forceZenModeWall?: boolean;
-      };
-   };
-   _environmentalRemoval?: string[];
-}