+++ types/beatmap/shared/custom/mod.ts
@@ -1,6 +1,6 @@
 export * from './bpm.ts';
 export * from './chroma.ts';
+export * from './contributor.ts';
 export * from './customData.ts';
-export * from './editor.ts';
 export * from './heck.ts';
 export * from './noodleExtensions.ts';